Implemented part of ticket #395 Pre-reordering commit
git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@512 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ gui/dialogs/variableviewer.cpp
@@ -27,7 +27,7 @@
 #define CLASSCOLUMN 2
 #define VALUECOLUMN 3
 
-std::string toString(int nNum);
+std::string toString(size_t nNum);
 
 extern Language _guilang;
 using namespace wxcode;
